refactored unit tests into separate directories within test directories
