+++ app/src/test/java/com/example/nostack/modeltests/EventTest.java
@@ -1,4 +1,4 @@
-package com.example.nostack;
+package com.example.nostack.modeltests;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
